Total drag coefficient

Total drag coefficient

Usage

C_t(
  u_c,
  u_b,
  macro_state = c(biomass, hm),
  SA_WW = 0.5 * (0.0306/2),
  site_params,
  constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)

Arguments

macro_state

vector of named numbers. Must include:

biomass, macroalgae wet weight (g)

hm, algae height (m)

site_params

vector of named numbers. Must include:

hz, total water depth (m)

hc, vertical water column occupied by the canopy (m)

d_top, depth of the top of the canopy beneath the water surface (m)

constants

vector of named numbers defining extra constants for the attenuation submodel. Must include:

s = 0.0045

gam = 1.13

a2 = 0.2^2

Cb = 0.0025

Value

The total drag coefficient C_t