Plot a species uptake rate over a range of N concentrations

Plot a species uptake rate over a range of N concentrations

Usage

N_plot(
  spec_params,
  uptake_shape,
  abbreviations = c(ni = "nitrate", am = "ammonium", ot = "other")
)

Arguments

spec_params

a vector of named numbers. Must include:

M and C for linear uptake, OR

V and K for Michaelis-Menton uptake

uptake_shape of "linear" or "MM" (or "Michaelis-Menton"). Not required if only one set of the above parameters are provided.

uptake_shape

kinetic shape for substrate uptake. One of "linear" or "Michaelis-Menton" (or "MM"). Defaults to Michaelis-Menton.

conc

a vector of the range of N concentrations to be tested (defaults to seq(0, 50, 0.25))

Value

a ggplot object of nitrogen uptake rates

Examples

examples
#> Error: object 'examples' not found