Salinity limitation on growth

Given species parameters, returns the relative limitation on growth rate according to a CTMI curve: $$ S_{lim} &=& \frac{(Sal-S_{max})(Sal-S_{min})^2}{(S_{opt}-S_{min})[(S_{opt}-S_{min})(Sal-S_{opt})-(S_{opt}-S_{max})(S_{opt}+S_{min}-2Sal)]} $$ The CTMI curve was formulated for temperature.

Usage

S_lim(Sal, spec_params)

Arguments

Sal

salinity to evaluate (g/L)

spec_params

a vector of named numbers. Must include:

S_opt the optimum salinity for macroalgae growth

S_min the minimum salinity for macroalgae growth (when Sal <= S_min, growth = 0)

S_max the maximum salinity for macroalgae growth (when Sal >= S_max, growth = 0)

Value

a scalar of relative salinity limitation on growth (between 0 and 1)

Examples

my_seaweed <- c(S_opt = 20, S_min = 5, S_max = 30)

S_lim(Sal = 22, spec_params = my_seaweed)
#> [1] 0.9633333

S_range <- 1:30
sapply(S_range, S_lim, spec_params = my_seaweed)
#>  [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02000000
#>  [7] 0.07215686 0.14666667 0.23578947 0.33333333 0.43428571 0.53454545
#> [13] 0.63072464 0.72000000 0.80000000 0.86871795 0.92444444 0.96571429
#> [19] 0.99126437 1.00000000 0.99096774 0.96333333 0.91636364 0.84941176
#> [25] 0.76190476 0.65333333 0.52324324 0.37122807 0.19692308 0.00000000