Temperature limitation on growth

Given species parameters, returns the relative limitation on growth rate according to a CTMI curve: $$\begin{array}[ccc] T_{lim} &=& \frac{(T_c-T_{max})(T_c-T_{min})^2}{(T_{opt}-T_{min})[(T_{opt}-T_{min})(T_c-T_{opt})-(T_{opt}-T_{max})(T_{opt}+T_{min}-2T_c)]} \end{array}$$

Usage

T_lim(Tc, spec_params)

Arguments

Tc

temperature to evaluate

spec_params

a vector of named numbers. Must include:

T_opt the optimum temperature for macroalgae growth

T_min the minimum temperature for macroalgae growth (when T_c < T_min, growth = 0)

T_max the maximum temperature for macroalgae growth (when T_c > T_max, growth = 0)

Value

a scalar of relative temperature limitation on growth (between 0 and 1)

Examples

my_seaweed <- c(T_opt = 20, T_min = 5, T_max = 30)

T_lim(Tc = 22, spec_params = my_seaweed)
#> [1] 0.9633333

T_range <- 1:30
sapply(T_range, T_lim, spec_params = my_seaweed)
#>  [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.02000000
#>  [7] 0.07215686 0.14666667 0.23578947 0.33333333 0.43428571 0.53454545
#> [13] 0.63072464 0.72000000 0.80000000 0.86871795 0.92444444 0.96571429
#> [19] 0.99126437 1.00000000 0.99096774 0.96333333 0.91636364 0.84941176
#> [25] 0.76190476 0.65333333 0.52324324 0.37122807 0.19692308 0.00000000