Asparagopsis armata model parameters

A named vector containing physiological and morphological parameters for the macroalgae species Asparagopsis armata used in the growth model.

Usage

a_armata

Format

A named numeric vector with 34 elements:

V_am

Maximum ammonium uptake rate (Michaelis-Menton uptake, mg gDW-1 d-1)

K_am

Half-saturation constant for ammonium uptake (Michaelis-Menton uptake, mg m-3)

M_am

Ammonium uptake rate slope (linear uptake, m3 gDW-1 d-1)

C_am

Ammonium uptake rate constant (linear uptake, m3 gDW-1 d-1)

V_ni

Maximum nitrate uptake rate (Michaelis-Menton uptake, mg gDW-1 d-1)

K_ni

Half-saturation constant for nitrate uptake (Michaelis-Menton uptake, mg m-3)

M_ni

Nitrate uptake rate slope (linear uptake, m3 gDW-1 d-1)

C_ni

Nitrate uptake rate constant (linear uptake, m3 gDW-1 d-1)

V_ot

Maximum other nitrogen uptake rate (Michaelis-Menton uptake, mg gDW-1 d-1)

K_ot

Half-saturation constant for other nitrogen uptake (Michaelis-Menton uptake, mg m-3)

M_ot

Other nitrogen uptake rate slope (linear uptake, m3 gDW-1 d-1)

C_ot

Other nitrogen uptake rate constant (linear uptake, m3 gDW-1 d-1)

Q_min

Minimum nitrogen quotient (mg gDW-1)

Q_max

Maximum nitrogen quotient (mg gDW-1)

K_c

Half-saturation constant for growth (mg gDW-1)

mu

Maximum growth rate (d-1)

D_m

Base loss rate (d-1)

D_ve

Loss rate with velocity (m-1 s d-1)

D_lo

Loss rate at low turbidity (d-1)

D_mi

Loss rate at medium turbidity (d-1)

D_hi

Loss rate at high turbidity (d-1)

a_cs

Carbon-specific self-shading constant (m mg-1)

I_o

Light saturation parameter (μmol photons m-2 s-1)

T_opt

Optimal temperature (°C)

T_min

Minimum temperature (°C)

T_max

Maximum temperature (°C)

S_opt

Optimal salinity (g L-1)

S_min

Minimum salinity (g L-1)

S_max

Maximum salinity (g L-1)

h_a

Height-controlling parameter a (d-1)

h_b

Height-controlling parameter b (d-1)

h_c

Height-controlling parameter c (d-1)

h_max

Maximum height (m)

DWWW

Dry weight to wet weight ratio (gWW gDW-1)

Source

Parameters derived from literature values and used in Asparagopsis armata growth modeling here (insert citation).

See also

Examples

# View all parameters
a_armata
#>         V_am         K_am         M_am         C_am         V_ni         K_ni 
#> 4.991988e+00 1.402071e+02 3.360000e-03 2.954853e+00           NA           NA 
#>         M_ni         C_ni        Q_min        Q_max          K_c           mu 
#> 0.000000e+00 1.138183e-01 1.943162e+01 4.303730e+01 6.000000e+00 3.472222e-01 
#>          D_m         D_ve         D_lo         D_mi         D_hi         a_cs 
#>           NA 1.533634e-02 5.389931e-02 1.270484e-01 2.567621e-01 2.509013e-06 
#>          I_o        T_opt        T_min        T_max        S_opt        S_min 
#> 1.500000e+02 1.686568e+01 8.895487e+00 2.242924e+01 3.525000e+01 3.375000e+01 
#>        S_max          h_a          h_b          h_c        h_max         DWWW 
#> 3.625000e+01 3.850000e+03 1.750000e+00 1.000000e-03 2.600000e-01 7.680000e+00 

# Access specific parameters
a_armata["mu"]  # Maximum growth rate
#>        mu 
#> 0.3472222 
a_armata["h_max"]  # Maximum height
#> h_max 
#>  0.26