Pre-check all parameters for grow_macroalgae()

Check that all parameters, inputs and settings are correct for the grow_macroalgae() function. Gives a report on what needs to be modified for the main function to run smoothly. This is to avoid the main function slowing down to give endless warnings and messages.

Usage

check_grow(
  t = 1:30,
  temperature,
  salinity,
  light,
  kW,
  velocity,
  nitrate,
  ammonium,
  ni_uptake,
  am_uptake,
  site_params,
  spec_params,
  initials,
  sparse_output = T,
  other_constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)

Arguments

temperature

a vector of daily temperatures (\(^{\circ}\)C)

salinity

a vector of daily salt concentrations (g L\(^{-1}\))

light

a vector of surface light (umol m\(^{-2}\) s\(^{-1}\))

kW

a vector of light attenuation coefficients for open water (m\(^{-1}\))

velocity

a vector of water velocities (m s\(^{-1}\))

nitrate

a vector of nitrate concentrations (mg m\(^{-3}\))

ammonium

a vector of ammonium concentrations (mg m\(^{-3}\))

ni_uptake

shape for nitrate uptake, either "MM" for Michaelis-Menton saturation kinetics or "linear" for linear uptake

am_uptake

shape for ammonium uptake, either "MM" for Michaelis-Menton saturation kinetics or "linear" for linear uptake

site_params

a named vector of species-specific parameters - see details

spec_params

a named vector of site-specific parameters - see details

initials

a named vector of the macroalgae starting conditions

sparse_output

logical, whether to include input vectors and other non-essential information in final dataframe (default = TRUE)

other_constants

a named vector of other constants for intermediate equations (see u_c())

Value

printout of potential errors for main function

Details

Example csv with all the spec_params & site_params required?

Examples