Determine limiting factor in growth

For each timestep in the growth output matrix, determine the factor(s) most likely to be limiting growth. This function is unfinished and subject to change.

Usage

get_limiting(output, spec_params)

Arguments

output

matrix, direct output of 'grow_macroalgae' function (can be sparse)

spec_params

a named vector of site-specific parameters - same as input to 'grow_macroalgae' function

Value

vector of limiting factors at each timestep

Details

This function gives an idea of what factor(s) are most likely to be limiting growth at each timestep. The choices are listed in order below. For each limitation, it can be assumed that the factors above it are within "ideal" ranges

T_lim: The macroalgae is limited by temperature

I_lim: The macroalgae is limited by light availability

S_lim: The macroalgae is limited by salinity

Q_lim: The macroalgae does not have enough stored nutrients to fix into biomass at maximum rate allowed by temperature, light and salinity

Ns_to_Nf: The macroalgae does not have enough stored nutrients to fix into biomass at maximum rate allowed by temperature, light and salinity

Nf_loss: More Nf was lost than was fixed

conc_nitrate/conc_ammonium/conc_other: There were not enough ambient nutrients available to replace fixed Ns

Ns_fixed_not_replaced: the Ns fixed last timestep was not replaced this timestep (Q_rel is declining)

Ns_loss_not_replaced: the Ns lost last timestep was not replaced this timestep (Q_rel is declining)

See also