Get the correct uptake rate

Get the correct uptake rate

Usage

get_uptake(conc, uptake_shape, Nform_abbr, spec_params)

Arguments

conc

substrate concentration

uptake_shape

kinetic shape for substrate uptake. One of "linear" or "Michaelis-Menton" (or "MM"). Defaults to Michaelis-Menton.

Nform_abbr

the abbreviation used in spec_params for the relevant nitrogen form. See details.

spec_params

a vector of named numbers. Must include:

M and C for linear uptake, OR

V and K for Michaelis-Menton uptake

uptake_shape of "linear" or "MM" (or "Michaelis-Menton"). Not required if only one set of the above parameters are provided.

Details

Gives the substrate uptake rate based on the shape and species-specific parameters supplied. If no shape is supplied, attempts to infer the correct shape from supplied parameters.

Nform_abbr allows the function to link species parameters with concentration. E.g. if Nform_abbr = "amm" (for ammonium) the function will look for M_amm and C_amm or V_amm and K_amm in spec_params and will ignore other uptake parameters which may be included for other substrates.

See also