Grow macroalgae

Initiate and grow macroalgae.

Usage

grow_macroalgae(
  start,
  grow_days,
  temperature,
  salinity,
  light,
  velocity,
  nitrate,
  ammonium,
  other_N,
  ni_uptake,
  am_uptake,
  site_params,
  spec_params,
  initials,
  other_constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)

Arguments

start

date, start of the growth period, date of at-sea deployment

grow_days

integer, number of day in growing period - if missing will take the length of the temperature vector

temperature

a vector of daily temperatures (C)

light

a vector of incoming light (umol m-2 s-1)

velocity

a vector of water velocities (m s-1)

nitrate

a vector of nitrate concentrations (mg m-3)

ammonium

a vector of ammonium concentrations (mg m-3)

other_N

a vector of other nitrogen concentrations (mg N m-3)

site_params

a named vector of species-specific parameters - see details

spec_params

a named vector of site-specific parameters - see details

initials

a named vector of the macroalgae starting conditions

other_constants

a named vector of miscellaneous constants, including:

s = 0.0045, the experimentally-derived shape factor for macroalgae movement in water

gam = 1.13, the modified drag coefficient for macroalgae

a2 = 0.2^2,

Cb = 0.0025, the bottom friction coefficient

Value

dataframe of outputs

Details

Example csv with all the spec_params & site_params required?

Note that the final growth dataframe is inclusive of the start and end date, so the environmental vectors must be the same

Examples