Plot of species growth response over a range of temperatures

Plot of species growth response over a range of temperatures

Usage

plot_T_range(T_range = seq(-10, 35, 0.25), spec_params)

Arguments

T_range

a vector of the range of temperatures to be tested (defaults to seq(-10, 35, 0.25))

spec_params

a vector of named numbers. Must include:

T_opt the optimum temperature for macroalgae growth

T_min the minimum temperature for macroalgae growth (when T_c < T_min, growth = 0)

T_max the maximum temperature for macroalgae growth (when T_c > T_max, growth = 0)

Value

a ggplot object of growth response over the specified temperature range

See also

Examples

examples
#> Error: object 'examples' not found