Relative water attenuation within canopy

#' @description A short description...

Usage

u_c(
  U0,
  macro_state,
  site_params,
  spec_params,
  constants = c(s = 0.0045, gam = 1.13, a2 = 0.2^2, Cb = 0.0025)
)

Arguments

U0

incoming incident water velocity (m/s)

macro_state

vector of named numbers. Must include:

biomass, macroalgae wet weight (g)

hm, algae height (m)

site_params

vector of named numbers. Must include:

hz, total water depth (m)

hc, vertical water column occupied by the canopy (m)

d_top, depth of the top of the canopy beneath the water surface (m)

spec_params

vector of named numbers. Must include:

SA_WW, conversion of wet weight to surface area

constants

vector of named numbers defining extra constants for the attenuation submodel. Must include all constants, which have the following default values:

s = 0.0045

gam = 1.13

a2 = 0.2^2

Cb = 0.0025

Value

the relative water attenuation coefficient (u_c)

See also